# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MetaRadar",
    "version": "5.1.0"
  },
  "paths": {
...
    "/api/v1/confluence": {
      "get": {
        "tags": [
          "Intelligence Views"
        ],
        "summary": "Get Confluence Alerts",
        "description": "Retrieve multi-source intelligence confluence alerts with real computed scoring and breakdown.",
        "operationId": "get_confluence_alerts_api_v1_confluence_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConfluenceAlertItem"
                  },
                  "title": "Response Get Confluence Alerts Api V1 Confluence Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/confluence/{confluence_id}/inspect": {
      "get": {
        "tags": [
          "Intelligence Views"
        ],
        "summary": "Inspect Confluence",
        "description": "Backward Trace Inspectability: Answers 'Why this confluence score?' with full verbatim citations,\nsource URLs, retrieval timestamps, and exact mathematical breakdown.",
        "operationId": "inspect_confluence_api_v1_confluence__confluence_id__inspect_get",
        "parameters": [
          {
            "name": "confluence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Confluence Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfluenceInspectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lifecycles": {
      "get": {
        "tags": [
          "Intelligence Views"
        ],
        "summary": "Get Lifecycle Timelines",
        "description": "Retrieve chronological lifecycle timeline events for developments and drug assets.",
        "operationId": "get_lifecycle_timelines_api_v1_lifecycles_get",
        "parameters": [
          {
            "name": "disease",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Disease"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LifecycleTimelineItem"
                  },
                  "title": "Response Get Lifecycle Timelines Api V1 Lifecycles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/red-team": {
      "get": {
        "tags": [
          "Intelligence Views"
        ],
        "summary": "Get Red Team Contradictions",
        "description": "Retrieve pairwise Red-Team contradiction alerts with verbatim evidence excerpts.",
        "operationId": "get_red_team_contradictions_api_v1_red_team_get",
        "parameters": [
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContradictionItem"
                  },
                  "title": "Response Get Red Team Contradictions Api V1 Red Team Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/missing-signals": {
      "get": {
        "tags": [
          "Intelligence Views"
        ],
        "summary": "Get Missing Signals",
        "description": "Retrieve missing signal watch items with explicit 6-state FSM and overdue heuristic scores.",
        "operationId": "get_missing_signals_api_v1_missing_signals_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MissingSignalWatchItem"
                  },
                  "title": "Response Get Missing Signals Api V1 Missing Signals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfluenceAlertItem": {
        "properties": {
          "confluence_id": {
            "type": "string",
            "format": "uuid",
            "title": "Confluence Id"
          },
          "development_id": {
            "type": "string",
            "format": "uuid",
            "title": "Development Id"
          },
          "development_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Development Title"
          },
          "signal_count": {
            "type": "integer",
            "title": "Signal Count"
          },
          "confluence_type": {
            "type": "string",
            "title": "Confluence Type"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "signals": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Signals"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "calculation_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calculation Version",
            "default": "confluence_v2.0"
          },
          "independent_sources_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Independent Sources Count"
          },
          "score_breakdown": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Breakdown"
          },
          "reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning"
          },
          "evidence_sources": {
            "items": {
              "$ref": "#/components/schemas/ConfluenceEvidenceSourceItem"
            },
            "type": "array",
            "title": "Evidence Sources"
          }
        },
        "type": "object",
        "required": [
          "confluence_id",
          "development_id",
          "signal_count",
          "confluence_type",
          "created_at"
        ],
        "title": "ConfluenceAlertItem"
      },
...
      "ConfluenceInspectResponse": {
        "properties": {
          "confluence_id": {
            "type": "string",
            "format": "uuid",
            "title": "Confluence Id"
          },
          "development_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Development Id"
          },
          "development_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Development Title"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "confluence_type": {
            "type": "string",
            "title": "Confluence Type"
          },
          "window_hours": {
            "type": "integer",
            "title": "Window Hours",
            "default": 48
          },
          "distinct_sources_count": {
            "type": "integer",
            "title": "Distinct Sources Count"
          },
          "score_breakdown": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Score Breakdown"
          },
          "reasoning": {
            "type": "string",
            "title": "Reasoning"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/ConfluenceEvidenceSourceItem"
            },
            "type": "array",
            "title": "Sources"
          },
          "detected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Detected At"
          }
        },
        "type": "object",
        "required": [
          "confluence_id",
          "score",
          "label",
          "confluence_type",
          "distinct_sources_count",
          "reasoning",
          "detected_at"
        ],
        "title": "ConfluenceInspectResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LifecycleTimelineItem": {
        "properties": {
          "lifecycle_id": {
            "type": "string",
            "format": "uuid",
            "title": "Lifecycle Id"
          },
          "development_id": {
            "type": "string",
            "format": "uuid",
            "title": "Development Id"
          },
          "development_title": {
            "type": "string",
            "title": "Development Title"
          },
          "disease": {
            "type": "string",
            "title": "Disease"
          },
          "asset_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Name"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "event_date": {
            "type": "string",
            "format": "date-time",
            "title": "Event Date"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "lifecycle_id",
          "development_id",
          "development_title",
          "disease",
          "stage",
          "event_date"
        ],
        "title": "LifecycleTimelineItem"
      },
      "MissingSignalWatchItem": {
        "properties": {
          "watch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Watch Id"
          },
          "development_id": {
            "type": "string",
            "format": "uuid",
            "title": "Development Id"
          },
          "development_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Development Title"
          },
          "trigger_event": {
            "type": "string",
            "title": "Trigger Event"
          },
          "expected_event": {
            "type": "string",
            "title": "Expected Event"
          },
          "monitoring_window_days": {
            "type": "integer",
            "title": "Monitoring Window Days"
          },
          "responsible_function": {
            "type": "string",
            "title": "Responsible Function"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 0.5
          },
          "confidence_type": {
            "type": "string",
            "title": "Confidence Type",
            "default": "overdue_heuristic"
          },
          "overdue_heuristic_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overdue Heuristic Score"
          },
          "days_overdue": {
            "type": "integer",
            "title": "Days Overdue",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "watch_id",
          "development_id",
          "trigger_event",
          "expected_event",
          "monitoring_window_days",
          "responsible_function",
          "status",
          "created_at"
        ],
        "title": "MissingSignalWatchItem"
      },
...
    }
  }
}